+++ m_near.h
@@ -1042,9 +1042,9 @@ extern int8_t   menu_messageString[105];
 extern int16_t  messageLastMenuActive;
 extern boolean  messageNeedsInput;
 extern void     (__near *messageRoutine)(int16_t response);
-extern int16_t  saveStringEnter;
-extern int16_t  saveSlot;       // which slot to save in
-extern int16_t  saveCharIndex;  // which char we're editing
+extern int8_t   saveStringEnter;
+extern int8_t   saveSlot;       // which slot to save in
+extern int16_t   saveCharIndex;  // which char we're editing
 extern int8_t   saveOldString[SAVESTRINGSIZE];
 extern int16_t  itemOn;                 // menu item skull is on
